requirements.txt
apache-airflow
apache-airflow-providers-microsoft-azure
pandas
numpy
pyarrow